+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * Result accept twice run
 * Message not append a file to post
+* HTTP reponse other than 200 is not handled
 
 ## Version 3.12.6 (2020-05-15)